test(ShadowServer): Move tests
