+++ external-import/shadowserver/tests/.env.test
@@ -1,0 +1,18 @@
+# OpenCTI configuration
+OPENCTI_URL="http://test-opencti-url"
+OPENCTI_TOKEN="test-opencti-token"
+
+# Common connector configuration
+CONNECTOR_ID="ChangeMe"
+CONNECTOR_NAME="Shadowserver .env.test"
+CONNECTOR_SCOPE="stix2"
+CONNECTOR_LOG_LEVEL="info"
+CONNECTOR_DURATION_PERIOD="PT5M"
+
+# Shadowserver configuration
+SHADOWSERVER_API_KEY="CHANGEME"
+SHADOWSERVER_API_SECRET="CHANGEME"
+SHADOWSERVER_MARKING="TLP:CLEAR"
+SHADOWSERVER_CREATE_INCIDENT="true"
+SHADOWSERVER_INCIDENT_SEVERITY="high"
+SHADOWSERVER_INCIDENT_PRIORITY="P1"
