--- external-import/shadowserver/tests/lib/.env.test
@@ -1,10 +1,0 @@
-# OpenCTI configuration
-OPENCTI_URL="http://test-opencti-url"
-OPENCTI_TOKEN="test-opencti-token"
-
-# Common connector configuration
-CONNECTOR_ID="test-connector-id"
-CONNECTOR_NAME="External Import Connector Template"
-CONNECTOR_SCOPE="ChangeMe"
-CONNECTOR_LOG_LEVEL="info"
-CONNECTOR_DURATION_PERIOD="PT5M"
