--- external-import/shadowserver/tests/lib/config.test.yml
@@ -1,10 +1,0 @@
-opencti:
-  url: "http://test-opencti-url"
-  token: "test-opencti-token"
-
-connector:
-  id: "test-connector-id"
-  name: "External Import Connector Template"
-  scope: "ChangeMe"
-  log_level: "info"
-  duration_period: "PT5M"
