+++ external-import/shadowserver/tests/test-requirements.txt
@@ -1,3 +1,3 @@
 -r ../requirements.txt
-pytest==8.3.5
-pytest-mock==3.14.0
+pytest~=8.3.5
+pytest-mock~=3.14.0
